reworked loginController present logic to modal view instead of initial view to get rid of a glitch on a startup if the user already authorised in app + moved some supporting data to utils dir
