+++ InstaVK/InstaVK/AppDelegate.swift
@@ -10,6 +10,8 @@ import UIKit
 import CoreData
 import VK_ios_sdk
 
+
+
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
@@ -17,6 +19,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+        VKSdk.initialize(withAppId: applicationID) // Initialize VK SDK with our app id (5986161)
         // Override point for customization after application launch.
         return true
     }
