+++ InstaVK/InstaVK/VkAppUtils.swift
@@ -1,0 +1,11 @@
+//
+//  VkAppUtils.swift
+//  InstaVK
+//
+//  Created by Svyatoslav Bykov on 23.04.17.
+//  Copyright © 2017 Nikita Susoev. All rights reserved.
+//
+
+import Foundation
+
+let applicationID =  "5986161"
